% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.r
\name{block_wise}
\alias{block_wise}
\title{<brief desc>
perform blockwise operation by applying function FUN to x and y
<full description>}
\usage{
block_wise(x, y, size_of_block, ncore, file_name, path, FUN, vmode = "single",
  ...)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{size_of_block}{size of a block final matrix will be split into}

\item{ncore}{number of cores to use for block wise operations}

\item{file_name}{name of the file to hold final ff matrix}

\item{path}{directory where the ff matrix will be stored}

\item{FUN}{function to apply}

\item{vmode}{= "single" storage mode for the final ff matrix}

\item{...}{other parameters, will be passed to FUN}
}
\value{
ff matrix
}
\description{
<brief desc>
perform blockwise operation by applying function FUN to x and y
<full description>
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link[ff]{ff}}
}

